move get validator count to /rpc/prysm/validator
